feat: refactor methods
