+++ formatter/src/test/kotlin/edu/austral/ingsis/gradle/formatter/FormatterTest.kt
@@ -16,7 +16,6 @@ import edu.austral.ingsis.gradle.common.token.Position
 import edu.austral.ingsis.gradle.common.token.Token
 import edu.austral.ingsis.gradle.common.token.TypeNumber
 import edu.austral.ingsis.gradle.common.token.ValueNumber
-import edu.austral.ingsis.gradle.formatter.util.Rule
 import edu.austral.ingsis.gradle.formatter.util.RuleParser
 import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
